Use lazy getSubscriptionManager for React 18 reusable state
React 18 introduces a strict mode behavior on initial mount which
unmounts and remounts the component with persisted state. In other
words, useEffect and useEffect cleanup functions run, but render and
memoize functions do not.

Relevant docs:
https://reactjs.org/docs/strict-mode.html#ensuring-reusable-state
https://github.com/reactwg/react-18/discussions/18

This causes document and view subscriptions to be prematurely
unsubscribed in React 18 when rendered in strict mode.

The recommendation from the React team is to switch to a lazy
initialization function for side-effects to be shared with other
components. This PR changes Provider to expose a lazy
getSubscriptionManager function to be used by other hooks' effects.